PUBDEV-7138 - first version of Python interface for Extended Isolation Forest
Python "from h2o.estimators import H2OExtendedIsolationForestEstimator" is working properly.
Extension Level parameter is working properly but validation is missing because I am not able to override init method in ExtendedIsolationForest.java with working rest API.
API is broken when the method init() is overriden. I was force to change sampleSize and extensionLevel from cammel case to snake because of the API generation method use reflection and naming is important.
Same reason to change sample_size to _sample_size (leading underscore). There is still a lot of unused parameters in ExtendedIsolationForestModelV3 that need to be solved.

+++ h2o-algos/src/main/java/hex/tree/isoforextended/SubSampleTask.java
@@ -13,12 +13,12 @@ import java.util.concurrent.atomic.AtomicInteger;
  */
 public class SubSampleTask extends MRTask<SubSampleTask> {
 
-    private int subSampleSize;
+    private long subSampleSize;
     private Random random;
     
     private AtomicInteger currentSubSampleSize = new AtomicInteger(-1);
 
-    public SubSampleTask(int subSampleSize, long seed) {
+    public SubSampleTask(long subSampleSize, long seed) {
         this.subSampleSize = subSampleSize;
         this.random = RandomUtils.getRNG(seed);
     }
